Moved utility methods to util.
