+++ librato_python_web/instrumentor/web/django_.py
@@ -26,7 +26,8 @@ import time
 from math import floor
 
 from librato_python_web.instrumentor.instrument import instrument_methods, function_wrapper_factory, \
-    generator_wrapper_factory, unwrap_method
+    generator_wrapper_factory
+from instrumentor.util import unwrap_method
 from librato_python_web.instrumentor import context as context
 from librato_python_web.instrumentor import telemetry
 from librato_python_web.instrumentor.telemetry import generate_record_telemetry
